pdf2csv/requirements.txt
pandas
numpy
tabula-py
